requirements/requirements_dev.txt
aiohttp==3.7.4.post0
asgiref==3.3.4
async-timeout==3.0.1
attrs==21.2.0
bootstrap4==0.1.0
cachetools==4.2.2
certifi==2021.5.30
chardet==4.0.0
Django==3.2.4
django-countries==7.2.1
django-crispy-forms==1.12.0
django-debug-toolbar==3.2.1
django-user-agents==0.4.0
djangorestframework==3.12.4
idna==2.10
ipinfo==4.2.0
multidict==5.1.0
Pillow==8.2.0
psycopg2==2.9.1
pytz==2021.1
requests==2.25.1
sqlparse==0.4.1
typing-extensions==3.10.0.0
ua-parser==0.10.0
urllib3==1.26.5
user-agents==2.2.0
yarl==1.6.3
